Moved calculating real cave width to services
